Rename module file names
